LPS-136653 Rename camel case package
